requirements.txt
pathway[xpack-llm-local,xpack-llm-docs]==0.9.0
pathway-llm-app
sentence-transformers==2.2.2
ollama
openai
pandas
python-dotenv
